Refactor UseCase
