+++ frontend/app/src/test/java/com/example/frontend/LoginUIKtTest.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.compose.runtime.mutableStateOf
 import com.example.frontend.api.AuthService
 import com.example.frontend.model.AuthResponse
-import com.example.frontend.usecase.loginUseCase
+import com.example.frontend.usecase.LoginUseCase
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -39,7 +39,7 @@ class LoginUIKtTest {
     fun loginButtonHandler() {
         val result = mutableStateOf("")
 
-        loginUseCase(mockContext, "test@example.com", "password", result, mockAuthAPI)
+        LoginUseCase(mockContext, "test@example.com", "password", result, mockAuthAPI).execute()
 
         // Capture the callback
         val argumentCaptor =
